+++ code/go/0chain.net/smartcontract/storagesc/blobber.go
@@ -2,7 +2,6 @@ package storagesc
 
 import (
 	"0chain.net/smartcontract/partitions"
-	"0chain.net/smartcontract/utils"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -495,7 +494,7 @@ func (sc *StorageSmartContract) commitBlobberRead(t *transaction.Transaction,
 	details.ReadReward += value // stat
 	details.Spent += value      // reduce min lock demand left
 
-	rewardRound := utils.GetCurrentRewardRound(balances.GetBlock().Round, conf.BlockReward.TriggerPeriod)
+	rewardRound := GetCurrentRewardRound(balances.GetBlock().Round, conf.BlockReward.TriggerPeriod)
 
 	if blobber.LastRoundDataReadUpdated >= rewardRound {
 		blobber.DataReadLastRound += sizeRead
@@ -776,7 +775,7 @@ func (sc *StorageSmartContract) commitBlobberConnection(
 			"moving tokens: %v", err)
 	}
 
-	startRound := utils.GetCurrentRewardRound(balances.GetBlock().Round, conf.BlockReward.TriggerPeriod)
+	startRound := GetCurrentRewardRound(balances.GetBlock().Round, conf.BlockReward.TriggerPeriod)
 
 	if blobber.RewardPartition.StartRound >= startRound && blobber.RewardPartition.Timestamp > 0 {
 		part, err := getOngoingPassedBlobbersList(balances, conf.BlockReward.TriggerPeriod)
