--- code/go/0chain.net/smartcontract/utils/utils.go
@@ -1,22 +1,0 @@
-package utils
-
-func GetCurrentRewardRound(currentRound, period int64) int64 {
-	if period > 0 {
-		extra := currentRound % period
-		if currentRound >= extra {
-			return currentRound - extra
-		} else {
-			return 0
-		}
-	}
-	return 0
-}
-
-func GetPreviousRewardRound(currentRound, period int64) int64 {
-	crr := GetCurrentRewardRound(currentRound, period)
-	if crr >= period {
-		return crr - period
-	} else {
-		return 0
-	}
-}
